requirements.txt
gym>=0.20.0
numpy>=1.21.2
pillow>=8.3.2
pyglet>=1.5.21
setuptools>=58.0.4
pymunk>=6.2.0
